requirements.txt
line-bot-sdk==3.7
flask==3.0.0
google-generativeai
requests
git+https://github.com/ozgur/python-firebase
